requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
celery==5.3.4
python-dotenv==1.0.0
djangorestframework==3.14.0
social-auth-app-django==5.3.0
openai==1.3.0
twilio==8.10.0
django-oauth-toolkit==2.3.0
redis==5.0.1
